Restructure files to include the new paths to the folder
