+++ src/components/ResourcesHeader.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core";
 
-import turtleResourceImage from "../assets/artem@3x.png";
+import turtleResourceImage from "../assets/resourcesPage/artem@3x.png";
 
 import SharedHeader from "./SharedHeader";
